properly convert lagtime + separate functions off
